# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is bioos apiserver using Hertz.",
        "title": "BioOS Apiserver",
        "contact": {
            "name": "hertz-contrib",
            "url": "https://github.com/hertz-contrib"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/.well-known/configuration": {
            "get": {
                "description": "get client configuration",
                "produces": [
                    "application/json"
                ],
                "summary": "use to get client configuration",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiserver.clientConfig"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "internal system error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
        "/workspace": {
            "get": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "description": "list workspaces",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "use to list workspaces",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "query page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "query size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query order, just like field1,field2:desc",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query searchWord",
                        "name": "searchWord",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "query exact",
                        "name": "exact",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query ids, split by comma",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ListWorkspacesResponse"
                        }
                    },
                    "400": {
                        "description": "invalid param",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "internal system error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "description": "import workspace",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "use to import workspace",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "workspace mount path",
                        "name": "mountType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "workspace mount type, only support nfs",
                        "name": "mountPath",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ImportWorkspaceResponse"
                        }
                    },
                    "400": {
                        "description": "invalid param",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "internal system error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "description": "create workspace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "use to create workspace",
                "parameters": [
                    {
                        "description": "create workspace request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateWorkspaceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateWorkspaceResponse"
                        }
                    },
                    "400": {
                        "description": "invalid param",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "internal system error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiserver.clientConfig": {
            "type": "object",
            "properties": {
                "notebook": {
                    "type": "object",
                    "properties": {
                        "officialImages": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/notebook.Image"
                            }
                        },
                        "resourceOptions": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/notebook.ResourceSize"
                            }
                        }
                    }
                },
                "storage": {
                    "type": "object",
                    "properties": {
                        "fsPath": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CreateWorkspaceRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "storage": {
                    "$ref": "#/definitions/handlers.WorkspaceStorage"
                }
            }
        },
        "handlers.CreateWorkspaceResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ImportWorkspaceResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ListWorkspacesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.WorkspaceItem"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "handlers.NFSWorkspaceStorage": {
            "type": "object",
            "properties": {
                "mountPath": {
                    "type": "string"
                }
            }
        },
...
        "handlers.WorkspaceItem": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "storage": {
                    "$ref": "#/definitions/handlers.WorkspaceStorage"
                },
                "updateTime": {
                    "type": "integer"
                }
            }
        },
        "handlers.WorkspaceStorage": {
            "type": "object",
            "properties": {
                "nfs": {
                    "$ref": "#/definitions/handlers.NFSWorkspaceStorage"
                }
            }
        },
...
        "notebook.GPU": {
            "type": "object",
            "properties": {
                "card": {
                    "description": "float is for mgpu",
                    "type": "number"
                },
                "memory": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                }
            }
        },
...
        "notebook.Image": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "notebook.ResourceSize": {
            "type": "object",
            "properties": {
                "cpu": {
                    "type": "number"
                },
                "disk": {
                    "type": "integer"
                },
                "gpu": {
                    "$ref": "#/definitions/notebook.GPU"
                },
                "memory": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "basicAuth": {
            "type": "basic"
        }
    }
}